pyproject.toml
[project]
name = "agentic-graph-rag"
version = "0.1.0"
description = "Agentic Graph RAG system with dynamic epistemic reasoning for quantitative finance research"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.60",
    "langchain>=0.3.15",
    "langchain-neo4j>=0.2.0",
    "google-genai>=1.0.0",
    "llama-index>=0.12.0",
    "llama-index-llms-gemini>=0.4.0",
    "llama-index-embeddings-gemini>=0.3.0",
    "neo4j>=5.26.0",
    "networkx>=3.4.2",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pypdf>=5.1.0",
    "python-dotenv>=1.0.1",
    "click>=8.1.7",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
]

[project.scripts]
graph-rag = "src.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
